Restructuring RDI home component folders to make way for new RDI statistics page
