+++ redisinsight/ui/src/pages/rdi/home/empty-message/EmptyMessage.spec.tsx
