+++ redisinsight/ui/src/pages/rdi/home/empty-message/EmptyMessage.tsx
