+++ redisinsight/ui/src/pages/rdi/home/empty-message/styles.module.scss
