+++ redisinsight/ui/src/pages/rdi/home/header/RdiHeader.tsx
