+++ redisinsight/ui/src/pages/rdi/home/instance-list/RdiInstancesListWrapper.tsx
