+++ redisinsight/ui/src/pages/rdi/home/instance-list/styles.module.scss
