+++ redisinsight/ui/src/pages/rdi/home/search/SearchRdiList.spec.tsx
