+++ redisinsight/ui/src/pages/rdi/home/search/styles.module.scss
